refactor(responses): isolate payload and stream mechanics
